# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NexusLab API",
    "description": "Marketplace API for B2B and B2C connections. Includes authentication, products, shops, B2B connections, leads, messaging, wallet, advertising, and AI features.",
    "version": "1.0.0",
    "contact": {
      "name": "NexusLab Support",
      "email": "api@nexuslab.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost/api",
      "description": "Local development"
    },
    {
      "url": "https://api.nexuslab.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List all products",
        "parameters": [
          {"name": "category", "in": "query", "schema": {"type": "integer"}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      }
    },
    "/shops": {
      "get": {
        "tags": ["Shops"],
        "summary": "List all shops",
        "parameters": [
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "city", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "List of shops"
          }
        }
      }
    },
...
  }
}